Attempt to fix the environment
For debugging purpose, it would be nice to fix the environment. The current issue is the env.seed() method only fix the random generator class within the gym environment, and it does not fix the numpy seed. I added a line that fix the numpy seed, and it would be a temporary solution.
The problem with this code is that all the numpy random generator would be fixed. If we want to use numpy random somewhere else along with fixed environment, it would cause some problems. Ideally, all the random number generation in initializing environment needs to use the random generator class with its own seed.